requirements.txt
streamlit==1.35.0
python-chess==1.999
requests==2.31.0
scikit-learn==1.3.0
hdbscan==0.8.33
numpy